Port vm problems to i64

+++ 2019/src/day00.rs
@@ -2,7 +2,7 @@ use aoc_runner_derive::{aoc, aoc_generator};
 use intcode::vm::Vm;
 
 #[aoc_generator(dayN)]
-pub fn parse_input(input: &str) -> Vec<i32> {
+pub fn parse_input(input: &str) -> Vec<i64> {
     input
         .lines()
         .map(|line| line.trim().parse().unwrap())
@@ -10,6 +10,6 @@ pub fn parse_input(input: &str) -> Vec<i32> {
 }
 
 #[aoc(dayN, part1)]
-pub fn part1(input: &[i32]) -> i32 {
+pub fn part1(input: &[i64]) -> i64 {
     unimplemented!();
 }
